pyproject.toml
[project]
name = "pdf-organizer"
version = "0.1.0"
description = "Organize PDF books into folders by topic via keyword classification."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pypdf>=4.0.0",
  "pdfminer.six>=20231228",
  "PyYAML>=6.0.1",
  "rich>=13.0.0",
]

[project.scripts]
pdf-organizer = "pdf_organizer.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
